pyproject.toml
[tool.poetry]
name = "explore-llm-text-classification"
version = "0.1.0"
description = "Demo possible llm api interactions"
authors = ["tab1tha"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-decouple = "^3.8"
openai = "^1.61.1"
google-generativeai = "^0.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
